+++ app/src/main/java/com/jeremiascortes/flowguide/features/settings/domain/model/SettingsState.kt
@@ -1,0 +1,7 @@
+package com.jeremiascortes.flowguide.features.settings.domain.model
+
+data class SettingsState(
+
+
+    val isLoading: Boolean = false
+)