+++ app/src/main/java/com/jeremiascortes/flowguide/features/settings/presentation/screen/SettingScreen.kt
@@ -1,0 +1,48 @@
+package com.jeremiascortes.flowguide.features.settings.presentation.screen
+
+import androidx.compose.material3.Button
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.collectAsState
+import com.jeremiascortes.flowguide.features.auth.presentation.AuthViewModel
+import com.jeremiascortes.flowguide.features.auth.presentation.NavigationEvent
+import com.jeremiascortes.flowguide.features.settings.presentation.viewmodel.SettingsViewModel
+import com.jeremiascortes.flowguide.presentation.components.AutoCollapsableTopAppBar
+
+@Composable
+fun SettingScreen(
+    settingsViewModel: SettingsViewModel,
+    authViewModel: AuthViewModel,
+    onLogout: () -> Unit,
+    onBack: () -> Unit
+) {
+    val state = settingsViewModel.state.collectAsState()
+
+    // Observar eventos de navegación
+    LaunchedEffect(Unit) {
+        authViewModel.navigationEvent.collect { event ->
+            when (event) {
+                is NavigationEvent.ToLogin -> onLogout()
+                is NavigationEvent.ToHome -> { /* no aplica aquí */ }
+            }
+        }
+    }
+
+    AutoCollapsableTopAppBar(
+        title = { Text(text = "Settings") },
+        onBack = onBack,
+    ) {
+        item {
+            Button(onClick = {
+                authViewModel.logout()
+            }){
+                Text(text = if (state.value.isLoading) "Cerrando sesión..." else "Cerrar Sesión")
+            }
+        }
+
+        items(40){
+            Text(text = "Item $it")
+        }
+    }
+}