+++ app/src/main/java/com/jeremiascortes/flowguide/navigation/AppNavHost.kt
@@ -34,23 +34,56 @@ import androidx.compose.animation.fadeIn
 import androidx.compose.animation.fadeOut
 import androidx.compose.animation.slideInHorizontally
 import androidx.compose.animation.slideOutHorizontally
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Check
+import androidx.compose.material.icons.filled.Edit
+import androidx.compose.material.icons.filled.Settings
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
+import com.jeremiascortes.flowguide.features.auth.domain.model.AuthState
 import com.jeremiascortes.flowguide.features.auth.presentation.AuthViewModel
 import com.jeremiascortes.flowguide.features.auth.presentation.LoginScreen
 import com.jeremiascortes.flowguide.features.auth.presentation.RegisterScreen
 import com.jeremiascortes.flowguide.features.home.presentation.screens.HomeScreen
 import com.jeremiascortes.flowguide.features.home.presentation.viewmodel.HomeViewModel
-import com.jeremiascortes.flowguide.features.procedure.presentation.screens.ProcedureScreen
+import com.jeremiascortes.flowguide.features.procedure.presentation.screen.ProcedureScreen
 import com.jeremiascortes.flowguide.features.procedure.presentation.viewmodel.ProcedureViewModel
+import com.jeremiascortes.flowguide.features.settings.presentation.screen.SettingScreen
+import com.jeremiascortes.flowguide.features.settings.presentation.viewmodel.SettingsViewModel
 import com.jeremiascortes.flowguide.features.welcome.presentation.SplashScreen
+import com.jeremiascortes.flowguide.presentation.components.BottomBar.BottomBar
+import com.jeremiascortes.flowguide.presentation.components.BottomBar.BottomBarItem
 
 @Composable
-fun Navigation() {
+fun AppNavHost() {
     val navController = rememberNavController()
+
+
+    val bottomBarItems = remember {
+        listOf(
+            BottomBarItem(
+                icon = Icons.Filled.Check,
+                contentDescription = "Completar",
+                onClick = { /* acción al hacer click */ }
+            ),
+            BottomBarItem(
+                icon = Icons.Filled.Edit,
+                contentDescription = "Editar",
+                onClick = { /* acción al hacer click */ }
+            ),
+            BottomBarItem(
+                icon = Icons.Filled.Settings,
+                contentDescription = "Configuración",
+                onClick = { navController.navigate(AppRoute.Settings.General) }
+            )
+        )
+    }
 
     NavHost(
         navController = navController,
@@ -84,12 +117,23 @@ fun Navigation() {
         // ==================== AUTH FLOW ====================
 
         composable<AppRoute.Auth.Splash> {
+            val authViewModel: AuthViewModel = hiltViewModel()
+            val authState by authViewModel.authState.collectAsState()
+
             SplashScreen(
-                onNavigateToHome = {
-                    navController.navigate(AppRoute.Auth.Login) {
-                        // Limpiar back stack para no volver al Splash
-                        popUpTo(AppRoute.Auth.Splash) { inclusive = true }
+                onNavigate = {
+                    if (authState is AuthState.Authenticated) {
+                        navController.navigate(AppRoute.Main.Home) {
+                            // Limpiar back stack para no volver al Splash
+                            popUpTo(AppRoute.Auth.Splash) { inclusive = true }
+                        }
+                    } else {
+                        navController.navigate(AppRoute.Auth.Login) {
+                            // Limpiar back stack para no volver al Splash
+                            popUpTo(AppRoute.Auth.Splash) { inclusive = true }
+                        }
                     }
+
                 }
             )
         }
@@ -135,7 +179,8 @@ fun Navigation() {
                 homeViewModel = homeViewModel,
                 onNavigateToProcedure = { idProcedure ->
                     navController.navigate(AppRoute.Main.Procedure(idProcedure)) {}
-                }
+                },
+                bottomBar = { BottomBar(items = bottomBarItems) }
             )
         }
 
@@ -144,6 +189,25 @@ fun Navigation() {
             ProcedureScreen(
                 procedureViewModel = viewModel,
                 onBack = { navController.popBackStack() }
+            )
+        }
+
+        composable<AppRoute.Settings.General> {
+            val settingsViewModel: SettingsViewModel = hiltViewModel()
+            val authViewModel: AuthViewModel = hiltViewModel()
+
+            SettingScreen(
+                settingsViewModel = settingsViewModel,
+                authViewModel = authViewModel,
+                onLogout = {
+                    // Limpiar back stack para no volver a Home
+                    navController.navigate(AppRoute.Auth.Splash) {
+                        popUpTo(AppRoute.Settings.General) { inclusive = true }
+                    }
+                },
+                onBack = {
+                    navController.popBackStack()
+                }
             )
         }
     }