+++ app/src/main/java/com/jeremiascortes/flowguide/presentation/components/BottomBar/BottomBarItem.kt
@@ -1,0 +1,13 @@
+package com.jeremiascortes.flowguide.presentation.components.BottomBar
+
+import androidx.compose.ui.graphics.vector.ImageVector
+
+/**
+ * Modelo que representa un item de acción en el BottomBar.
+ */
+data class BottomBarItem(
+    val icon: ImageVector,
+    val contentDescription: String,
+    val onClick: () -> Unit,
+    val enabled: Boolean = true
+)