+++ app/src/main/java/com/jeremiascortes/flowguide/presentation/components/MainTopBar.kt
@@ -1,0 +1,62 @@
+package com.jeremiascortes.flowguide.presentation.components
+
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Favorite
+import androidx.compose.material.icons.filled.Menu
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.PlainTooltip
+import androidx.compose.material3.Text
+import androidx.compose.material3.TooltipAnchorPosition
+import androidx.compose.material3.TooltipBox
+import androidx.compose.material3.TooltipDefaults
+import androidx.compose.material3.TopAppBar
+import androidx.compose.material3.rememberTooltipState
+import androidx.compose.runtime.Composable
+
+/**
+ * A sample for a simple use of small [TopAppBar].
+ *
+ * The top app bar here does not react to any scroll events in the content under it.
+ */
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun MainTopBar(
+    title: @Composable () -> Unit,
+) {
+    TopAppBar(
+        title = title,
+        navigationIcon = {
+            TooltipBox(
+                positionProvider =
+                    TooltipDefaults.rememberTooltipPositionProvider(
+                        TooltipAnchorPosition.Above
+                    ),
+                tooltip = { PlainTooltip { Text("Menu") } },
+                state = rememberTooltipState(),
+            ) {
+                IconButton(onClick = { /* doSomething() */ }) {
+                    Icon(imageVector = Icons.Filled.Menu, contentDescription = "Menu")
+                }
+            }
+        },
+        actions = {
+            TooltipBox(
+                positionProvider =
+                    TooltipDefaults.rememberTooltipPositionProvider(
+                        TooltipAnchorPosition.Above
+                    ),
+                tooltip = { PlainTooltip { Text("Add to favorites") } },
+                state = rememberTooltipState(),
+            ) {
+                IconButton(onClick = { /* doSomething() */ }) {
+                    Icon(
+                        imageVector = Icons.Filled.Favorite,
+                        contentDescription = "Add to favorites",
+                    )
+                }
+            }
+        },
+    )
+}